+++ biosppy/ml/__init__.py
@@ -1,1 +1,1 @@
-from . import ecg, utils
+from . import ecg_ml, utils_ml
